requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
paho-mqtt==1.6.1
python-socketio==5.10.0
python-engineio==4.8.0
